setting "dtype=torch.float64"

+++ python/exp05_explicit_inversion.py
@@ -17,14 +17,14 @@ def solve_rec(A, B):
 
 def exp05_explicit_inversion(b, n):
 
-    A = torch.randn(n, n)
-    B = torch.randn(n, 200)
+    A = torch.randn((n, n), dtype=torch.float64)
+    B = torch.randn((n, 200), dtype=torch.float64)
 
     res1 = b.benchmark('solve_nai', solve_nai, A, B)
     res2 = b.benchmark('solve_rec', solve_rec, A, B)
     logger.info('Solve correctness: {}'.format(torch.allclose(res1, res2)))
 
-    B = torch.randn(n, 5*n)  # B.rows , 5 * A.rows
+    B = torch.randn((n, 5*n), dtype=torch.float64)  # B.rows , 5 * A.rows
     res3 = b.benchmark('solve_large_nai', solve_nai, A, B)
     res4 = b.benchmark('solve_large_rec', solve_rec, A, B)
     logger.info('Solve correctness: {}'.format(torch.allclose(res3, res4)))
